+++ vite.config.ts
@@ -1,4 +1,4 @@
-import { defineConfig } from "vitest/config";
+import { defineConfig, configDefaults } from "vitest/config";
 import react from "@vitejs/plugin-react";
 import cssInjectedByJsPlugin from "vite-plugin-css-injected-by-js";
 
@@ -27,5 +27,15 @@ export default defineConfig({
     globals: true,
     environment: "jsdom",
     setupFiles: ["./test/init.ts"],
+    coverage: {
+      exclude: [
+        ...configDefaults.coverage.exclude!,
+        "docs/**",
+        "examples/**",
+        "src/lib.tsx",
+        "src/main.tsx",
+        "src/themes/**",
+      ],
+    },
   },
 });